Migrate 'linkedin stats' to serverless job (#295)
* Init Linkedin stats job

* Add Terraform
