+++ api/package.json
@@ -81,7 +81,6 @@
     "supertest": "^7.1.1",
     "ts-node": "10.9.2",
     "typescript": "5.8.3",
-    "vitest": "3.1.4",
-    "@types/he": "^1.2.3"
+    "vitest": "3.1.4"
   }
 }
