+++ process/src/index.ts
@@ -25,7 +25,6 @@ import { captureException } from "./error";
 import brevo from "./jobs/brevo";
 import imports from "./jobs/import";
 import leboncoin from "./jobs/leboncoin";
-import linkedinStats from "./jobs/linkedin-stats";
 import metabase from "./jobs/metabase";
 import moderation from "./jobs/moderation";
 import report from "./jobs/report";
@@ -34,7 +33,6 @@ const app = express();
 
 const runnings = {
   mission: false,
-  linkedinStats: false,
   metabase: false,
   leboncoin: false,
   report: false,
@@ -161,45 +159,6 @@ const metabaseJob = new CronJob(
   "Europe/Paris"
 );
 
-// Every friday at 09:00 AM
-const linkedinStatsJob = new CronJob(
-  "0 9 * * 5",
-  async () => {
-    const checkInId = Sentry.captureCheckIn({
-      monitorSlug: "linkedin-stats",
-      status: "in_progress",
-    });
-    if (ENVIRONMENT !== "production") {
-      Sentry.captureCheckIn({
-        checkInId,
-        monitorSlug: "linkedin-stats",
-        status: "ok",
-      });
-      return;
-    }
-    runnings.linkedinStats = true;
-    try {
-      await linkedinStats.handler();
-      Sentry.captureCheckIn({
-        checkInId,
-        monitorSlug: "linkedin-stats",
-        status: "ok",
-      });
-    } catch (error) {
-      captureException(error);
-      Sentry.captureCheckIn({
-        checkInId,
-        monitorSlug: "linkedin-stats",
-        status: "error",
-      });
-    }
-    runnings.linkedinStats = false;
-  },
-  null,
-  true,
-  "Europe/Paris"
-);
-
 // Every day at 10:00 AM
 const leboncoinJob = new CronJob(
   "0 10 * * *",
@@ -275,14 +234,6 @@ app.get("/tasks", async (req, res) => {
         lastRun: missionJob.lastDate(),
         nextRun: missionJob.nextDate(),
         running: runnings.mission,
-      },
-      {
-        name: "Update Linkedin Stats",
-        schedule: linkedinStatsJob.cronTime.source,
-        started: linkedinStatsJob.isActive,
-        lastRun: linkedinStatsJob.lastDate(),
-        nextRun: linkedinStatsJob.nextDate(),
-        running: runnings.linkedinStats,
       },
       {
         name: "Update Metabase",
